requirements.txt
# --------------------------
# Core Framework
# --------------------------
fastapi
uvicorn
# fastapi[standard]
# uvicorn[standard]

# --------------------------
# Database
# --------------------------
sqlmodel
sqlalchemy
asyncpg
# alembic

# --------------------------
# Authentication / Security
# --------------------------
# python-jose[cryptography]
# passlib[bcrypt]
python-jose
passlib
bcrypt

# --------------------------
# Validation / Settings
# --------------------------
pydantic
pydantic-settings
# python-dotenv

# --------------------------
# Utilities
# --------------------------
# python-multipart
# email-validator

# --------------------------
# Development / Testing
# --------------------------
# pytest
# httpx
